Abstracted end to end tests for ephemeral workflows to base class
